requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
python-multipart==0.0.6
pillow==10.1.0
opencv-python==4.8.1.78
pytesseract==0.3.10
brotli==1.1.0
pymupdf==1.24.11
brotli==1.1.0
paddlepaddle==2.6.1; platform_machine == "x86_64"
paddleocr==2.8.1; platform_machine == "x86_64"
boto3==1.34.0
redis==5.0.1
celery==5.3.4
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
httpx==0.25.2
python-dotenv==1.1.1
email-validator==2.1.1
requests==2.32.3
streamlit==1.35.0
pytest==8.3.3
pytest-asyncio==0.23.8
openai==1.51.2

